requirements.txt
uuid
apache-airflow-providers-apache-spark
